refactor: Score 클래스 삭제 및 Cards에 점수 로직 이동
